+++ test/unit/lib/socket.spec.js
@@ -1,7 +1,7 @@
-const get = require('../../lib/socket');
-const { socketToken: url } = require('../../lib/utils/urls');
+const get = require('../../../lib/socket');
+const { socketToken: url } = require('../../../lib/utils/urls');
 
-jest.mock('../../lib/credentials', () => ({
+jest.mock('../../../lib/credentials', () => ({
   setSocketToken: jest.fn(),
 }));
 
@@ -11,7 +11,7 @@ const resultMock = {
   },
 };
 
-jest.mock('../../lib/request', () => ({
+jest.mock('../../../lib/utils/request', () => ({
   get: jest.fn().mockResolvedValue({
     data: {
       socket_token: 'socket_token',
@@ -19,8 +19,8 @@ jest.mock('../../lib/request', () => ({
   }),
 }));
 
-const { setSocketToken } = require('../../lib/credentials');
-const { get: requestGet } = require('../../lib/request');
+const { setSocketToken } = require('../../../lib/credentials');
+const { get: requestGet } = require('../../../lib/utils/request');
 
 describe('Socket', () => {
   it('Should get socket token and save it', (done) => {
@@ -29,7 +29,7 @@ describe('Socket', () => {
       expect(setSocketToken).toBeCalled();
       expect(setSocketToken).toBeCalledWith(resultMock.data.socket_token);
       expect(requestGet).toBeCalled();
-      expect(requestGet).toBeCalledWith(url, {});
+      expect(requestGet).toBeCalledWith(url);
       done();
     });
   });